// src/app/types/types.ts
export interface Product {
  reviews: boolean
  id: string
  name: string
  description: string
  category: string
  subCategory: string
  price: number
  oldPrice: number
  availability: string
  image: string
  images: string[]
  colors: string[]
  sizeAvailability: string[]
  rating: number
  material: string
}